Moved OpenIdAuthenticatorConfig into its own assembly SequenceConfig
