--- Assets/SequenceSDK/Authentication/ScriptableObjects.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 9d01ded11cd74f2b8e76609f31a9e1db
-timeCreated: 1702653834
