+++ Assets/SequenceSDK/Authentication/SequenceAuthentication.asmdef
@@ -1,3 +1,16 @@
 {
-	"name": "SequenceAuthentication"
+    "name": "SequenceAuthentication",
+    "rootNamespace": "",
+    "references": [
+        "GUID:a35e3a53d4439435f8b36ed2c6158cd8"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
 }
