+++ Assets/SequenceSDK/Config/SequenceConfig.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "SequenceConfig"
+}
