+++ Assets/SequenceSDK/Config/SequenceConfig.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a35e3a53d4439435f8b36ed2c6158cd8
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
